requirements.txt
amqp==2.5.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
boto3==1.9.197
botocore==1.12.197
celery==4.3.0
Django==2.2.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-livereload-server==0.3.2
django-queryset-csv==1.0.2
django-storages==1.7.1
django-tables2==2.0.6
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
docutils==0.14
jmespath==0.9.4
kombu==4.6.3
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
redis==3.2.1
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
tornado==6.0.3
unicodecsv==0.14.1
urllib3==1.25.3
vine==1.3.0
